% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{Attendence data for World Cup football matches.}
\format{
a data frame with 22 observations
\describe{
  \item{Totalattendance}{The total number of people who attended matches}
  \item{Averageattendance}{The average attendance at each match}
  \item{Matches}{The number of matches contested}
  \item{WorldCup}{The host country and year}
}
}
\usage{
World_Cup
}
\description{
Attendence data for World Cup football matches.
}
\keyword{datasets}
