% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Population.R
\docType{data}
\name{World_Population}
\alias{World_Population}
\title{Population data for all UN countries from 1950 to 2020.}
\format{
a data frame with 235 observations
\describe{
  \item{CountryName}{The country whose population is being described}
  \item{1950}{The country's population as of July of 1950}
  \item{2020}{The country's population as of July of 2020}
}
}
\usage{
World_Population
}
\description{
Population data for all UN countries from 1950 to 2020.
}
\keyword{datasets}
